PhongHNT update statistic

+++ src/main/java/vn/codegym/projections/IMonthCostProjection.java
@@ -1,5 +1,5 @@
 package vn.codegym.projections;
 
 public interface IMonthCostProjection {
-    Integer getTotalCost();
+    Long getTotalCost();
 }
